[REFACTOR] #31 - 패키지 정리
