+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
 
         <variable
             name="viewModel"
-            type="com.jjsh.smartshopping.presentation.main.MainViewModel" />
+            type="com.jjsh.smartshopping.presentation.ui.main.MainViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
@@ -13,7 +13,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.main.MainActivity">
+        tools:context=".presentation.ui.main.MainActivity">
 
         <com.google.android.material.appbar.AppBarLayout
             android:id="@+id/abl_appbar_layout"
