+++ app/src/main/res/layout/activity_search.xml
@@ -5,7 +5,7 @@
 
         <variable
             name="viewModel"
-            type="com.jjsh.smartshopping.presentation.search.SearchViewModel" />
+            type="com.jjsh.smartshopping.presentation.ui.search.SearchViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
@@ -13,7 +13,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.search.SearchActivity">
+        tools:context=".presentation.ui.search.SearchActivity">
 
         <com.jjsh.smartshopping.presentation.customview.SearchTextView
             android:id="@+id/stv_search"
