+++ app/src/main/res/layout/fragment_cart.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background = "@color/purple_200"
-        tools:context=".presentation.main.cart.CartFragment">
+        tools:context=".presentation.ui.main.cart.CartFragment">
 
     </FrameLayout>
 </layout>
