+++ app/src/main/res/layout/fragment_chart.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/purple_500"
-        tools:context=".presentation.main.chart.ChartFragment">
+        tools:context=".presentation.ui.main.chart.ChartFragment">
 
     </FrameLayout>
 </layout>
