+++ app/src/main/res/layout/fragment_check_list.xml
@@ -6,7 +6,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/purple_700"
-        tools:context=".presentation.main.checklist.CheckListFragment">
+        tools:context=".presentation.ui.main.checklist.CheckListFragment">
 
     </FrameLayout>
 </layout>
