+++ app/src/main/res/layout/fragment_home.xml
@@ -7,14 +7,14 @@
 
         <variable
             name="viewModel"
-            type="com.jjsh.smartshopping.presentation.main.home.HomeViewModel" />
+            type="com.jjsh.smartshopping.presentation.ui.main.home.HomeViewModel" />
     </data>
 
     <androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.main.home.HomeFragment">
+        tools:context=".presentation.ui.main.home.HomeFragment">
 
         <com.google.android.material.appbar.AppBarLayout
             android:layout_width="match_parent"
