+++ app/src/main/res/layout/fragment_search_history.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.search.history.SearchHistoryFragment">
+        tools:context=".presentation.ui.search.history.SearchHistoryFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rv_search_history"
